update bbs.rb, etc.

+++ bbs.rb
@@ -1,5 +1,6 @@
 require 'sinatra'
 require 'active_record'
+require 'bcrypt'
 
 set :environment, :production
 
@@ -15,6 +16,7 @@ get '/' do
 end
 
 post '/' do
+  # treat username
   username = Rack::Utils.escape_html(params[:user_name])
   if username.empty?
     username = "No Name"
@@ -22,13 +24,24 @@ post '/' do
   if username.length > 30
     username = username[0,29] + "..."
   end
+
+  # encrypt password
+  delete_key = params[:del_key]
+  delete_key.present?
+    delete_key = 0000
+  end
+  password_salt = BCrypt::Engine.generate_salt
+  password_hash = BCrypt::Engine.hash_secret(delete_key, password_salt)
+
   message  = Rack::Utils.escape_html(params[:mess])
-  Post.create(message: message, posted_at: Time.now, user_name: username)
+  Post.create(message: message, posted_at: Time.now, user_name: username, hashed_delete_key: password_hash, salt: password_salt)
   redirect '/'
 end
 
 delete '/' do
   post = Post.find(params[:id])
-  post.destroy
+  if post && post.hashed_delete_key == BCrypt::Engine.hash_secret(params[:del_key], post.salt)
+    post.destroy
+  end
   redirect '/'
 end
